Code review changes. Refactor various scatterplot options.
