+++ src/components/scatterplot/ScatterplotOptions.js
@@ -13,7 +13,7 @@ import { GLSL_COLORMAPS } from '../../layers/constants';
 
 export default function ScatterplotOptions(props) {
   const {
-    customOptions,
+    children,
     observationsLabel,
     cellRadius,
     setCellRadius,
@@ -81,20 +81,9 @@ export default function ScatterplotOptions(props) {
     [handleColormapRangeChange],
   );
 
-  const customRows = customOptions ? customOptions.map(option => (
-    <TableRow key={`${option.label}-option-row`}>
-      <TableCell className={classes.labelCell}>
-        {option.label}
-      </TableCell>
-      <TableCell className={classes.inputCell}>
-        {option.input}
-      </TableCell>
-    </TableRow>
-  )) : [];
-
   return (
     <OptionsContainer>
-      {customRows}
+      {children}
       <CellColorEncodingOption
         observationsLabel={observationsLabel}
         cellColorEncoding={cellColorEncoding}
